Update .gitignore to exclude specific files
The .gitignore file has been updated to include new entries. Specifically, the FodyWeavers.xsd file and the appsettings1.json file located in the QuanLyQuanCafe.Server directory have been added to the ignore list. This means that these files will now be ignored by Git and will not be tracked or included in version control.